// artifacts/api-server/src/sya/lib/schemas.ts
export const ListProvisionsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
});
// ...
export const GetProvisionParams = zod.object({
  id: zod.coerce.number(),
});